requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-RESTX==1.3.0
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Cache
redis==5.0.1

# Validation
marshmallow==3.20.1

# Testing
pytest==7.4.3
hypothesis==6.92.1

# Server
gunicorn==21.2.0

# Environment
python-dotenv==1.0.0
